Actor Critic

Policy-based reinforcement learning control algorithm using both policy (the actor) and value function (the critic).

actorCritic(envir, fun.approx = "table", policy = "softmax",
  critic.type = "advantage", preprocessState = identity, n.episodes = 100,
  discount = 1, alpha = 0.01, beta = 0.1, lambda.actor = 0,
  lambda.critic = 0, updateAlpha = identity2, updateBeta = identity2,
  updateLambdaActor = identity2, updateLambdaCritic = identity2)

Arguments

envir
[R6 class] The reinforcement learning environment created by makeEnvironment.

fun.approx
[character(1)] Type of function approximator used for policy and value function. Currently supported are linear combination of features ("linear") and table lookup ("table").

policy
[character(1)] Policy type, supported are "softmax" for a discrete action space and "gaussian" for a continuous action space using a normal distribution.

critic.type
[character(1)] Type of the critic. Currently only an advantage actor critic is supported, which is approximated by the td error.

preprocessState
[function] A function that takes the state observation returned from the environment as an input and preprocesses this in a way the algorithm can work with it.

n.episodes
[integer(1)] Number of episodes.

discount
[numeric(1) in [0, 1]] Discount factor.

alpha
[numeric(1)] Learning rate (step size) for the policy.

beta
[numeric(1)] Learning rate (step size) for the critic.

lambda.actor
[numeric(1) in [0, 1]] Trace decay parameter for the actor.

lambda.critic
[numeric(1) in [0, 1]] Trace decay parameter for the critic.

updateAlpha
[function] A function which takes two arguments, alpha and the current number of episodes. Could be used to decay the parameter over time.

updateBeta
[function] A function which takes two arguments, beta and the current number of episodes. Could be used to decay the parameter over time.

updateLambdaActor
[function] A function which takes two arguments, lambda.actor and the current number of episodes. Could be used to decay the parameter over time.

updateLambdaCritic
[function] A function which takes two arguments, lambda.critic and the current number of episodes. Could be used to decay the parameter over time.

Value

[list] Returns a list with policy and value function parameters and some statistics about learning behaviour, e.g. the number of steps and return per episode.

Details

When using a gaussian policy mean and variance will be parametrized as a linear combination of features mu = x * w_1 and sigma = exp(x * w_2).

For a softmax policy the action preferences are computed by a linear combination of features.

Eligibility traces can be used for both policy parameters and value function parameters.

References

Sutton and Barto (Book draft 2017): Reinforcement Learning: An Introduction. Chapter 13

Examples

# Mountain Car
m = makeEnvironment("MountainCar")

# Define preprocessing function (we use grid tiling)
n.tilings = 8
max.size = 4096
iht = IHT(max.size)

position.max = m$state.space.bounds[[1]][2]
position.min = m$state.space.bounds[[1]][1]
velocity.max = m$state.space.bounds[[2]][2]
velocity.min = m$state.space.bounds[[2]][1]
position.scale = n.tilings / (position.max - position.min)
velocity.scale = n.tilings / (velocity.max - velocity.min)

# Scale state first, then get active tiles and return n hot vector
gridTiling = function(state) {
  state = c(position.scale * state[1], velocity.scale * state[2])
  active.tiles = tiles(iht, 8, state)
  nHot(active.tiles, max.size, out = "vector")
}

# Linear function approximation and softmax policy
res = actorCritic(m, fun.approx = "linear",
  preprocessState = gridTiling, n.episodes = 20)

#> Episode 1 finished after 1450 steps with a return of -1449

#> Episode 2 finished after 327 steps with a return of -326

#> Episode 3 finished after 290 steps with a return of -289

#> Episode 4 finished after 235 steps with a return of -234

#> Episode 5 finished after 233 steps with a return of -232

#> Episode 6 finished after 277 steps with a return of -276

#> Episode 7 finished after 236 steps with a return of -235

#> Episode 8 finished after 256 steps with a return of -255

#> Episode 9 finished after 155 steps with a return of -154

#> Episode 10 finished after 152 steps with a return of -151

#> Episode 11 finished after 187 steps with a return of -186

#> Episode 12 finished after 191 steps with a return of -190

#> Episode 13 finished after 157 steps with a return of -156

#> Episode 14 finished after 162 steps with a return of -161

#> Episode 15 finished after 204 steps with a return of -203

#> Episode 16 finished after 199 steps with a return of -198

#> Episode 17 finished after 153 steps with a return of -152

#> Episode 18 finished after 155 steps with a return of -154

#> Episode 19 finished after 151 steps with a return of -150

#> Episode 20 finished after 146 steps with a return of -145

#----------------
# Mountain Car with continuous action space
m = makeEnvironment("MountainCarContinuous")

# Linear function approximation and gaussian policy
set.seed(123)
res = actorCritic(m, fun.approx = "linear", policy = "gaussian",
  preprocessState = gridTiling, n.episodes = 20)

#> Episode 1 finished after 780 steps with a return of -779

#> Episode 2 finished after 319 steps with a return of -318

#> Episode 3 finished after 317 steps with a return of -316

#> Episode 4 finished after 157 steps with a return of -156

#> Episode 5 finished after 149 steps with a return of -148

#> Episode 6 finished after 194 steps with a return of -193

#> Episode 7 finished after 153 steps with a return of -152

#> Episode 8 finished after 132 steps with a return of -131

#> Episode 9 finished after 106 steps with a return of -105

#> Episode 10 finished after 151 steps with a return of -150

#> Episode 11 finished after 106 steps with a return of -105

#> Episode 12 finished after 136 steps with a return of -135

#> Episode 13 finished after 110 steps with a return of -109

#> Episode 14 finished after 98 steps with a return of -97

#> Episode 15 finished after 96 steps with a return of -95

#> Episode 16 finished after 101 steps with a return of -100

#> Episode 17 finished after 144 steps with a return of -143

#> Episode 18 finished after 96 steps with a return of -95

#> Episode 19 finished after 103 steps with a return of -102

#> Episode 20 finished after 165 steps with a return of -164