Multi-armed bandit

Different solution methods for multi-armed bandits.

bandit(rewardFun, n.actions, n.episodes = 100,
  action.selection = c("egreedy", "greedy", "UCB", "gradientbandit"),
  epsilon = 0.1, alpha = 0.1, initial.value = 0, initial.visits = 0,
  C = 2, updateEpsilon = identity2, updateAlpha = identity2)

Arguments

rewardFun
[function] A function, which takes an action (integer(1)) as first argument and returns a numeric scalar reward.

n.actions
[integer(1)] Number of actions.

n.episodes
[integer(1)] Number of episodes.

action.selection
[character(1)] Which method to use for action selection, one of "egreedy", "greedy", "UCB" or "gradientbandit".

epsilon
[numeric(1) in [0,1]] Ratio of random exploration in epsilon-greedy action selection.

alpha
[numeric(1)] Parameter of the gradient bandit algorithm, a higher alpha value gives more weight to recent rewards (useful for non-stationary environments).

initial.value
[numeric(1)] Initial values for the action values Q, set this to the maximal possible reward to encourage exploration (optimistic initialization).

initial.visits
[integer(1)] Set this to a high number to encourage exploration (together with a high initial.value).

C
[numeric(1)] Controls the degree of exploration. High C values lead to more exploration.

updateEpsilon
[function] A function which takes two arguments, epsilon and the current number of episodes. Could be used to decay the parameter over time.

updateAlpha
[function] A function which takes two arguments, alpha and the current number of episodes. Could be used to decay the parameter over time.

Value

[numeric] Returns the action values for the arms of the bandit or, for gradient-bandit action selection, the probabilities for each action.

Details

Multi-armed bandits are a simplified reinforcement learning problem, each arm of the bandit pays off a reward and the goal is to maximize this reward, i.e. to choose the best arm. The arms of the bandit can be seen as actions, after each action the episode ends (there are no states). To find the best action, the algorithm is faced with a tradeoff between exploration and exploitation.

Actions are numerated starting with 0!

Examples

set.seed(123)

# Define reward function
rewardFun = function(action) {
  if (action == 0) {
    reward = rnorm(1, mean = 1, sd = 1)
  }
  if (action == 1) {
    reward = rnorm(1, mean = 2, sd = 4)
  }
  if (action == 2) {
    reward = runif(1, min = 0, max = 5)
  }
  if (action == 3) {
    reward = rexp(1, rate = 0.25)
  }
  reward
}

bandit(rewardFun, n.actions = 4, n.episodes = 1000,
  action.selection = "greedy")

#> [1] 0.9938972 1.6475474 2.4775240 0.1263094

bandit(rewardFun, n.actions = 4, n.episodes = 1000,
  action.selection = "egreedy", epsilon = 0.5)

#> [1] 1.091109 1.810578 2.368114 3.973869

bandit(rewardFun, n.actions = 4, n.episodes = 1000,
  action.selection = "greedy",
  initial.value = 5, initial.visits = 100)

#> [1] 4.175454 4.132243 4.152457 4.208284

bandit(rewardFun, n.actions = 4, n.episodes = 1000,
  action.selection = "UCB", C = 2)

#> [1] -0.1064495  2.5144118  2.7020018  4.0096271

# true values: 1, 2, 2.5, 4

# Gradient bandit algorithm
bandit(rewardFun, n.actions = 4, n.episodes = 10000,
  action.selection = "gradientbandit", alpha = 0.1)

#> [1] 0.0000690247 0.0001050642 0.0003705861 0.9994553250