Dynamic Programming

These functions solve a Markov Decision Process using a model of the environment. evaluatePolicy evaluates a given policy, iteratePolicy and iterateValue can be used to find the optimal policy.

evaluatePolicy(envir, policy, v = NULL, q = NULL, discount = 1,
  precision = 1e-04, n.iter = NULL)

iteratePolicy(envir, policy = NULL, discount = 1, n.iter = NULL,
  precision.eval = 1e-04, n.iter.eval = NULL)

iterateValue(envir, v = NULL, q = NULL, discount = 1, precision = 1e-04,
  n.iter = NULL)

Arguments

envir
[R6 class] The reinforcement learning environment created by makeEnvironment.

policy
[matrix(n.states x n.actions)] A policy specified as a probability matrix (states x actions).

v
[numeric(n.states)] Initial state value function. Terminal states must have a value of 0!

q
[matrix(n.states x n.actions)] Initial action value function. Terminal states must have a value of 0!

discount
[numeric(1) in [0, 1]] Discount factor.

precision
[numeric(1)] Algorithm stops when improvement is smaller than precision.

n.iter
[integer(1)] Number of iterations. If supplied the precision argument will be ignored.

precision.eval
[numeric(1)] Policy evaluation stops when improvement is smaller than precision.

n.iter.eval
[integer(1)] Number of iterations per evaluation step.

Value

[list(3)] Returns the state value function [numeric], the action value function [matrix] and the policy [matrix].

Details

evaluatePolicy runs until the improvement in the state value function in two subsequent steps is smaller than the given precision in all states or if the specified number of iterations is exhausted.

Both iteratePolicy and iterateValue alternate between evaluating a policy and improving the current policy by acting greedily with respect to the current policy's value function. The difference between these two algorithms is that iteratePolicy evaluates the policy until some stop criterion is met, while iterateValue evaluates each policy only one step and then immediately improves upon the current policy.

When the policy argument is NULL the initial policy will be a uniform random policy. iteratePolicy stops if the policy does not change in two subsequent iterations or if the specified number of iterations is exhausted. For the policy evaluation step in policy iteration the same stop criteria mentioned above are applied. iterateValue runs until the improvement in the value function in two subsequent steps is smaller than the given precision in all states or if the specified number of iterations is exhausted.

References

Sutton and Barto (Book draft 2017): Reinforcement Learning: An Introduction. Chapter 4.

Examples

# Set up gridworld problem
env = makeEnvironment("Gridworld", shape = c(4, 4), goal.states = c(0, 15))

# Define uniform random policy, take each action with equal probability
random.policy = matrix(1 / env$n.actions, nrow = env$n.states,
  ncol = env$n.actions)

# Evaluate given policy
res = evaluatePolicy(env, random.policy, precision = 0.001)
print(round(matrix(res$v, ncol = 4, byrow = TRUE)))

#>      [,1] [,2] [,3] [,4]
#> [1,]    0  -14  -20  -22
#> [2,]  -14  -18  -20  -20
#> [3,]  -20  -20  -18  -14
#> [4,]  -22  -20  -14    0

# Find optimal policy using Policy Iteration
res = iteratePolicy(env)
print(round(matrix(res$v, ncol = 4, byrow = TRUE)))

#>      [,1] [,2] [,3] [,4]
#> [1,]    0   -1   -2   -3
#> [2,]   -1   -2   -3   -2
#> [3,]   -2   -3   -2   -1
#> [4,]   -3   -2   -1    0

# Find optimal policy using Value Iteration
res = iterateValue(env)
print(res$policy)

#>       [,1] [,2] [,3] [,4]
#>  [1,]    1    0    0    0
#>  [2,]    1    0    0    0
#>  [3,]    1    0    0    0
#>  [4,]    1    0    0    0
#>  [5,]    0    0    1    0
#>  [6,]    1    0    0    0
#>  [7,]    1    0    0    0
#>  [8,]    0    0    0    1
#>  [9,]    0    0    1    0
#> [10,]    1    0    0    0
#> [11,]    0    1    0    0
#> [12,]    0    0    0    1
#> [13,]    0    1    0    0
#> [14,]    0    1    0    0
#> [15,]    0    1    0    0
#> [16,]    1    0    0    0