Mountain Car

The classical mountain car problem for reinforcement learning.

MountainCar

MountainCarContinuous

Format

An object of class R6ClassGenerator of length 24.

Details

The classical Mountain Car task the action is one of {0, 1, 2}, in the continuous version the action is in [-1, 1].

Usage

makeEnvironment("MountainCar") makeEnvironment("MountainCarContinuous")

Methods

$step(action)

Take action in environment. Returns a list with state, reward, done.

$reset()

Resets the done flag of the environment and returns an initial state. Useful when starting a new episode.

$visualize()

Visualizes the environment (if there is a visualization function).

Examples

env = makeEnvironment("MountainCar")
env$reset()

#>            [,1] [,2]
#> [1,] -0.4520213    0

env$step(1L)

#> $state
#>           [,1]          [,2]
#> [1,] -0.452554 -0.0005327148
#> 
#> $reward
#> [1] -1
#> 
#> $done
#> [1] FALSE
#> 

env = makeEnvironment("MountainCarContinuous")
env$reset()

#>            [,1] [,2]
#> [1,] -0.5169617    0

env$step(0.62)

#> $state
#>            [,1]         [,2]
#> [1,] -0.5160814 0.0008802249
#> 
#> $reward
#> [1] -1
#> 
#> $done
#> [1] FALSE
#>