Q(sigma) / Q-Learning / Sarsa / Expected Sarsa

Value-based reinforcement learning control algorithms.

qSigma(envir, fun.approx = "table", preprocessState = identity,
  model = NULL, initial.value = NULL, n.states = NULL, n.episodes = 100,
  sigma = 1, target.policy = "egreedy", lambda = 0,
  eligibility.type = 0, learning.rate = 0.1, epsilon = 0.1,
  discount = 1, double.learning = FALSE, update.target.after = 1,
  replay.memory = NULL, replay.memory.size = 1, batch.size = 1,
  alpha = 0, theta = 0.01, updateEpsilon = identity2,
  updateSigma = identity2, updateLambda = identity2,
  updateAlpha = identity2, updateLearningRate = identity2)

qlearning(envir, fun.approx = "table", preprocessState = identity,
  model = NULL, initial.value = NULL, n.states = NULL, n.episodes = 100,
  lambda = 0, eligibility.type = 0, learning.rate = 0.1, epsilon = 0.1,
  discount = 1, double.learning = FALSE, update.target.after = 1,
  replay.memory = NULL, replay.memory.size = 1, batch.size = 1,
  alpha = 0, theta = 0.01, updateEpsilon = identity2,
  updateSigma = identity2, updateLambda = identity2,
  updateAlpha = identity2, updateLearningRate = identity2)

sarsa(envir, fun.approx = "table", preprocessState = identity,
  model = NULL, initial.value = NULL, n.states = NULL, n.episodes = 100,
  lambda = 0, eligibility.type = 0, learning.rate = 0.1, epsilon = 0.1,
  discount = 1, double.learning = FALSE, update.target.after = 1,
  replay.memory = NULL, replay.memory.size = 1, batch.size = 1,
  alpha = 0, theta = 0.01, updateEpsilon = identity2,
  updateSigma = identity2, updateLambda = identity2,
  updateAlpha = identity2, updateLearningRate = identity2)

expectedSarsa(envir, fun.approx = "table", preprocessState = identity,
  model = NULL, initial.value = NULL, n.states = NULL, n.episodes = 100,
  target.policy = "egreedy", lambda = 0, eligibility.type = 0,
  learning.rate = 0.1, epsilon = 0.1, discount = 1,
  double.learning = FALSE, update.target.after = 1, replay.memory = NULL,
  replay.memory.size = 1, batch.size = 1, alpha = 0, theta = 0.01,
  updateEpsilon = identity2, updateSigma = identity2,
  updateLambda = identity2, updateAlpha = identity2,
  updateLearningRate = identity2)

Arguments

envir
[R6 class] The reinforcement learning environment created by makeEnvironment.

fun.approx
[character(1)] How to represent the value function? Currently "table", "linear" and "neural.network" are supported.

preprocessState
[function] A function that takes the state observation returned from the environment as an input and preprocesses this in a way the algorithm can work with it.

model
[keras model] A neural network model specified using the keras package. See Details for more information.

initial.value
[numeric] Initial value function matrix or weight matrix. If NULL weights will be initialized to 0. Only used for tabular or linear function approximation.

n.states
[integer(1)] Number of states for tabular value function. Only needed if the state space is continuous, but will be transformed by preprocessState, so that a single integer value is returned. Else the number of states is accessed from the envir argument.

n.episodes
[integer(1)] Number of episodes.

sigma
[numeric(1) in [0, 1]] Parameter of the Q(sigma) algorithm. It controls if the temporal-difference target is equal to the sarsa target (for sigma = 1) or the expected sarsa target (for sigma = 0). For intermediate values of sigma a weighted mean between the two targets is used.

target.policy
[character(1)] Should the temporal-difference target be computed on-policy using the epsilon-greedy behavior policy (target.policy = "egreedy") or off-policy using a greedy policy in the expected sarsa part of the update (target.policy = "greedy")?

lambda
[numeric(1) in [0, 1]] Eligibility trace decay parameter.

eligibility.type
[numeric(1)] Type of eligibility trace, use eligibility.type = 1 for replacing traces, eligibility.type = 0 for accumulating traces or intermediate values for a mixture between both.

learning.rate
[numeric(1)] Learning rate used for gradient descent.

epsilon
[numeric(1) in [0, 1]] Ratio of random exploration in epsilon-greedy action selection.

discount
[numeric(1) in [0, 1]] Discount factor.

double.learning
[logical(1)] Should double learning be used?

update.target.after
[integer(1)] When using double learning the target network / table will be updated after update.target.after steps.

replay.memory
[list] Initial replay memory, which can be passed on. Each list element must be a list containing state, action, reward and next.state.

replay.memory.size
[integer(1)] Size of the replay memory. Only used if replay.memory is NULL. Then the replay memory will be initially filled with experiences following a uniform random policy.

batch.size
[integer(1)] Batch size, how many experiences are sampled from the replay memory at each step? Must be smaller than the size of the replay memory!

alpha
[numeric(1) in [0, 1]] If alpha = 0 sampling from replay memory will be uniform, otherwise observations with high temporal-difference error will be proportionally prioritized.

theta
[numeric(1) in (0, 1]] theta is a small positive constant that prevents the edge-case of transitions in the replay memory not being revisited once their TD error is zero.

updateEpsilon
[function] A function that updates epsilon. It takes two arguments, epsilon and the current number of episodes which are finished, and returns the new epsilon value.

updateSigma
[function] A function that updates sigma. It takes two arguments, sigma and the current number of episodes which are finished, and returns the new sigma value.

updateLambda
[function] A function that updates lambda. It takes two arguments, lambda and the current number of episodes which are finished, and returns the new lambda value.

updateAlpha
[function] A function that updates alpha. It takes two arguments, alpha and the current number of episodes which are finished, and returns the new alpha value.

updateLearningRate
[function] A function that updates the learning rate. It takes two arguments, learning.rate and the current number of episodes which are finished, and returns the new learning.rate value.

Value

[list(4)] Returns the action value function or model parameters [matrix] and the number of steps per episode [numeric] and the sum of all rewards in the episode [numeric].

Details

You must specify the reinforcement learning environment using the envir argument. It takes an R6 class created by makeEnvironment as input. See documentation there.

The algorithms can be used to find the optimal action value function using the principle of generalized policy iteration. They all learn online using temporal-difference learning. Q(sigma) subsumes the well known Q-Learning, Sarsa and Expected Sarsa algorithms as special cases. The default call qsigma() is exactly equivalent to Sarsa(0). A weighted mean between sarsa and expected sarsa updates can be used by varying the parameter sigma. When target.policy == "egreedy" the policy used to compute the expected sarsa is the epsilon-greedy policy used for action selection, when target.policy == "greedy" a greedy target policy will be used as in Q-Learning. See De Asis et al. (2017) for more details.

The functions qlearning and sarsa are there for convenience. They call the qSigma function with a special set of parameters.

When fun.approx == "table" the action value function will be represented using a table, a linear combination of features and a neural network can also be used. For a neural network you need to pass on a keras model via the model argument. This way it is possible to construct a Deep Q-Network (DQN).

The raw state observation returned from the environment can be preprocessed using the preprocessState function. This function takes the state observation as input and returns a preprocessed state which can be directly used by the function approximator. This is especially important when using function approximation.

Experience replay can be used by specifying a prefilled replay memory through the replay.memory argument or by specifying the length of the replay memory, which will then be filled with random experience. Using the defaults experiences are sampled uniformly, a proportional prioritization proposed by Schaul et al. (2016) can also be used by varying the parameters alpha and theta.

Using eligibility traces the one-step algorithms can be extended to multi-step algorithms. The parameter lambda controls the tradeoff between one-step and multi-step methods. Three different kinds of eligibility traces are implemented, accumulating, replacing and dutch traces. See Sutton and Barto (Book draft 2017, Chapter 12) for more details. This only works if no experience replay is used, i.e. replay.memory.size == 1 and for a tabular value function.

Double Learning can be used with all of the functions. Then two Q value functions are used, Q1 for action selection and Q2 for action evaluation. After a number of steps the weights of Q2 are replaced by the weights of Q1. See Hasselt et al. (2015) for details.

The hyperparameters epsilon, theta, sigma, alpha, lambda and learning.rate can be changed over time. Therefore pass on functions that return a new value of the hyperparameter. These updates will be applied after each episode. First argument of the function must be the parameter itself, the second argument is the current episode number.

References

De Asis et al. (2017): Multi-step Reinforcement Learning: A Unifying Algorithm

Hasselt et al. (2015): Deep Reinforcement Learning with Double Q-Learning

Mnih et al. (2013): Playing Atari with Deep Reinforcement Learning

Schaul et al. (2016): Prioritized Experience Replay

Sutton and Barto (Book draft 2017): Reinforcement Learning: An Introduction

Examples

# Solve Windy Gridworld
env = windyGridworld()

res = qSigma(env, sigma = 0.5, n.episodes = 20)

#> Episode 1 finished after 719 steps with a return of -719

#> Episode 2 finished after 1823 steps with a return of -1823

#> Episode 3 finished after 644 steps with a return of -644

#> Episode 4 finished after 586 steps with a return of -586

#> Episode 5 finished after 264 steps with a return of -264

#> Episode 6 finished after 389 steps with a return of -389

#> Episode 7 finished after 299 steps with a return of -299

#> Episode 8 finished after 96 steps with a return of -96

#> Episode 9 finished after 252 steps with a return of -252

#> Episode 10 finished after 223 steps with a return of -223

#> Episode 11 finished after 117 steps with a return of -117

#> Episode 12 finished after 301 steps with a return of -301

#> Episode 13 finished after 146 steps with a return of -146

#> Episode 14 finished after 92 steps with a return of -92

#> Episode 15 finished after 409 steps with a return of -409

#> Episode 16 finished after 161 steps with a return of -161

#> Episode 17 finished after 168 steps with a return of -168

#> Episode 18 finished after 69 steps with a return of -69

#> Episode 19 finished after 122 steps with a return of -122

#> Episode 20 finished after 151 steps with a return of -151

print(res$steps)

#>  [1]  719 1823  644  586  264  389  299   96  252  223  117  301  146   92  409
#> [16]  161  168   69  122  151

res = qlearning(env, n.episodes = 20)

#> Episode 1 finished after 883 steps with a return of -883

#> Episode 2 finished after 1547 steps with a return of -1547

#> Episode 3 finished after 1014 steps with a return of -1014

#> Episode 4 finished after 131 steps with a return of -131

#> Episode 5 finished after 582 steps with a return of -582

#> Episode 6 finished after 340 steps with a return of -340

#> Episode 7 finished after 225 steps with a return of -225

#> Episode 8 finished after 243 steps with a return of -243

#> Episode 9 finished after 96 steps with a return of -96

#> Episode 10 finished after 258 steps with a return of -258

#> Episode 11 finished after 97 steps with a return of -97

#> Episode 12 finished after 419 steps with a return of -419

#> Episode 13 finished after 107 steps with a return of -107

#> Episode 14 finished after 181 steps with a return of -181

#> Episode 15 finished after 91 steps with a return of -91

#> Episode 16 finished after 226 steps with a return of -226

#> Episode 17 finished after 222 steps with a return of -222

#> Episode 18 finished after 99 steps with a return of -99

#> Episode 19 finished after 104 steps with a return of -104

#> Episode 20 finished after 94 steps with a return of -94

print(matrix(round(apply(res$Q1, 1, max), 1), ncol = 10, byrow = TRUE))

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,] -4.2 -4.4 -4.9 -5.7 -6.3 -6.8 -6.8 -6.3 -5.5  -4.7
#> [2,] -4.0 -4.0 -4.2 -4.5 -4.2 -3.2 -1.9 -3.4 -3.8  -3.9
#> [3,] -3.7 -3.6 -3.5 -3.5 -2.6 -1.3 -0.5 -1.9 -2.7  -3.1
#> [4,] -3.3 -3.1 -2.9 -2.5 -1.4 -0.5  0.0  0.0 -2.0  -2.2
#> [5,] -2.8 -2.7 -2.3 -1.8 -0.8  0.0  0.0 -0.2 -0.8  -1.5
#> [6,] -2.5 -2.3 -1.8 -1.3 -0.3  0.0  0.0  0.0 -0.6  -0.9
#> [7,] -2.2 -2.0 -1.4 -0.7  0.0  0.0  0.0  0.0 -0.1  -0.5

# Decay epsilon over time. Each 10 episodes epsilon will be halfed.
decayEpsilon = function(epsilon, i) {
  if (i %% 10 == 0) {
    epsilon = epsilon * 0.5
  }
  epsilon
}

res = expectedSarsa(env, epsilon = 0.5, updateEpsilon = decayEpsilon, n.episodes = 20)

#> Episode 1 finished after 892 steps with a return of -892

#> Episode 2 finished after 2211 steps with a return of -2211

#> Episode 3 finished after 1916 steps with a return of -1916

#> Episode 4 finished after 481 steps with a return of -481

#> Episode 5 finished after 92 steps with a return of -92

#> Episode 6 finished after 313 steps with a return of -313

#> Episode 7 finished after 134 steps with a return of -134

#> Episode 8 finished after 451 steps with a return of -451

#> Episode 9 finished after 141 steps with a return of -141

#> Episode 10 finished after 141 steps with a return of -141

#> Episode 11 finished after 145 steps with a return of -145

#> Episode 12 finished after 28 steps with a return of -28

#> Episode 13 finished after 147 steps with a return of -147

#> Episode 14 finished after 21 steps with a return of -21

#> Episode 15 finished after 182 steps with a return of -182

#> Episode 16 finished after 92 steps with a return of -92

#> Episode 17 finished after 121 steps with a return of -121

#> Episode 18 finished after 100 steps with a return of -100

#> Episode 19 finished after 161 steps with a return of -161

#> Episode 20 finished after 83 steps with a return of -83

optimal.policy = max.col(res$Q1) - 1L
print(matrix(optimal.policy, ncol = 10, byrow = TRUE))

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    3    3    0    0    0    2    3    1    1     3
#> [2,]    1    3    3    3    0    3    0    1    3     3
#> [3,]    2    0    0    0    3    1    3    0    3     3
#> [4,]    1    2    3    3    1    3    3    0    1     3
#> [5,]    1    1    3    1    2    3    1    0    0     3
#> [6,]    3    3    1    3    3    2    0    2    3     0
#> [7,]    1    2    1    2    1    2    0    2    3     1

# Solve the Mountain Car problem using linear function approximation
m = mountainCar()

# Define preprocessing function (we use grid tiling)
n.tilings = 8
max.size = 4096
iht = IHT(max.size)

position.max = m$state.space.bounds[[1]][2]
position.min = m$state.space.bounds[[1]][1]
velocity.max = m$state.space.bounds[[2]][2]
velocity.min = m$state.space.bounds[[2]][1]
position.scale = n.tilings / (position.max - position.min)
velocity.scale = n.tilings / (velocity.max - velocity.min)

# Scale state first, then get active tiles and return n hot vector
gridTiling = function(state) {
  state = c(position.scale * state[1], velocity.scale * state[2])
  active.tiles = tiles(iht, 8, state)
  makeNHot(active.tiles, max.size, out = "vector")
}

set.seed(123)
res = sarsa(m, fun.approx = "linear", preprocessState = gridTiling, n.episodes = 20)

#> Episode 1 finished after 840 steps with a return of -839

#> Episode 2 finished after 784 steps with a return of -783

#> Episode 3 finished after 684 steps with a return of -683

#> Episode 4 finished after 694 steps with a return of -693

#> Episode 5 finished after 318 steps with a return of -317

#> Episode 6 finished after 475 steps with a return of -474

#> Episode 7 finished after 270 steps with a return of -269

#> Episode 8 finished after 274 steps with a return of -273

#> Episode 9 finished after 235 steps with a return of -234

#> Episode 10 finished after 199 steps with a return of -198

#> Episode 11 finished after 276 steps with a return of -275

#> Episode 12 finished after 160 steps with a return of -159

#> Episode 13 finished after 213 steps with a return of -212

#> Episode 14 finished after 304 steps with a return of -303

#> Episode 15 finished after 216 steps with a return of -215

#> Episode 16 finished after 235 steps with a return of -234

#> Episode 17 finished after 224 steps with a return of -223

#> Episode 18 finished after 180 steps with a return of -179

#> Episode 19 finished after 153 steps with a return of -152

#> Episode 20 finished after 222 steps with a return of -221

print(res$returns)

#>  [1] -839 -783 -683 -693 -317 -474 -269 -273 -234 -198 -275 -159 -212 -303 -215
#> [16] -234 -223 -179 -152 -221

# NOT RUN {
env = makeGridworld(c(4, 4), goal.states = 15, initial.state = 0)

# Use a neural network as function approximator
makeOneHot = function(state) {
  one.hot = matrix(rep(0, 16), nrow = 1)
  one.hot[1, state + 1] = 1
  one.hot
}

# Define keras model
library(keras)
model = keras_model_sequential()
model %>% layer_dense(units = 4, activation = 'linear', input_shape = c(16))

res = qSigma(env, fun.approx = "neural.network", model = model,
  preprocessState = makeOneHot, n.episodes = 20)
# }