TD(lambda)

Estimate the state value function for a given policy using temporal difference learning.

td(envir, policy, fun.approx = "table", preprocessState = identity,
  initial.value = NULL, n.episodes = NULL, n.steps = NULL, discount = 1,
  lambda = 0, eligibility.type = 0, learning.rate = 0.1,
  updateLambda = identity2, updateLearningRate = identity2)

Arguments

envir
[R6 class] The reinforcement learning environment created by makeEnvironment.

policy
[matrix | function] A policy specified as a probability matrix (states x actions) or a function returning an action given a preprocessed state observation (with linear function approximation).

fun.approx
[character(1)] How to represent the value function? Currently "table", "linear" and "neural.network" are supported.

preprocessState
[function] A function that takes the state observation returned from the environment as an input and preprocesses this in a way the algorithm can work with it.

initial.value
[numeric] Initial value function matrix or weight matrix. If NULL weights will be initialized to 0. Only used for tabular or linear function approximation.

n.episodes
[integer(1)] Number of episodes.

n.steps
[integer(1)] Number of evaluations (steps in the environment).

discount
[numeric(1) in [0, 1]] Discount factor.

lambda
[numeric(1) in [0, 1]] Eligibility trace decay parameter.

eligibility.type
[numeric(1)] Type of eligibility trace, use eligibility.type = 1 for replacing traces, eligibility.type = 0 for accumulating traces or intermediate values for a mixture between both.

learning.rate
[numeric(1)] Learning rate used for gradient descent.

updateLambda
[function] A function that updates lambda. It takes two arguments, lambda and the current number of episodes which are finished, and returns the new lambda value.

updateLearningRate
[function] A function that updates the learning rate. It takes two arguments, learning.rate and the current number of episodes which are finished, and returns the new learning.rate value.

Value

[list] Returns the state value function V and some statistics about learning behavior, e.g. the number of steps and return per episode.

Details

Can be used for episodic or continuing environments, please specify either a maximum number of steps or a maximum number of episodes.

References

Sutton and Barto (Book draft 2017): Reinforcement Learning: An Introduction. Chapter 6

Examples

# Random Walk Task (Sutton & Barto Example 6.2)
P = array(dim = c(7, 7, 2))
P[, , 1] = matrix(c(rep(c(1, rep(0, 6)), 2), c(0, 1, rep(0, 5)),
  c(0, 0, 1, rep(0, 4)), c(rep(0, 3), 1, rep(0, 3)), c(rep(0, 4), 1, rep(0, 2)),
  c(rep(0, 6), 1)), ncol = 7, byrow = TRUE)
P[, , 2] = matrix(c(c(1, rep(0, 6)), c(0, 0, 1, rep(0, 4)),
  c(rep(0, 3), 1, rep(0, 3)), c(rep(0, 4), 1, rep(0, 2)),
  c(rep(0, 5), 1, 0), c(rep(0, 6), 1), c(rep(0, 6), 1)), ncol = 7, byrow = TRUE)
R = matrix(c(rep(0, 12), 1, 0), ncol = 2)
env = makeEnvironment(transitions = P, rewards = R, initial.state = 3)

# Uniform random policy
random.policy = matrix(1 / env$n.actions, nrow = env$n.states,
  ncol = env$n.actions)

# Estimate state value function with TD(0)
res = td(env, random.policy, n.episodes = 20)

#> Episode 1 finished after 13 steps with a return of 1

#> Episode 2 finished after 35 steps with a return of 0

#> Episode 3 finished after 5 steps with a return of 0

#> Episode 4 finished after 3 steps with a return of 1

#> Episode 5 finished after 7 steps with a return of 1

#> Episode 6 finished after 7 steps with a return of 0

#> Episode 7 finished after 7 steps with a return of 0

#> Episode 8 finished after 21 steps with a return of 0

#> Episode 9 finished after 3 steps with a return of 0

#> Episode 10 finished after 15 steps with a return of 1

#> Episode 11 finished after 37 steps with a return of 0

#> Episode 12 finished after 5 steps with a return of 1

#> Episode 13 finished after 9 steps with a return of 0

#> Episode 14 finished after 9 steps with a return of 0

#> Episode 15 finished after 7 steps with a return of 1

#> Episode 16 finished after 9 steps with a return of 1

#> Episode 17 finished after 37 steps with a return of 1

#> Episode 18 finished after 3 steps with a return of 1

#> Episode 19 finished after 3 steps with a return of 1

#> Episode 20 finished after 3 steps with a return of 0