Actor Critic

Policy-based reinforcement learning control algorithm using both policy (the actor) and value function (the critic).

actorCritic(envir, fun.approx = "table", policy = "softmax",
  critic.type = "advantage", preprocessState = identity, n.episodes = 100,
  discount = 1, alpha = 0.01, beta = 0.1, lambda = 0)

Arguments

envir
[R6 class] The reinforcement learning environment created by makeEnvironment.

fun.approx
[character(1)] Type of function approximator used for policy and value function. Currently supported are linear combination of features ("linear") and table lookup ("table").

policy
[character(1)] Policy type, supported are "softmax" for a discrete action space and "gaussian" for a continuous action space using a normal distribution.

critic.type
[character(1)] Type of the critic. Currently only advantage actor critic is supported.

preprocessState
[function] A function that takes the state observation returned from the environment as an input and preprocesses this in a way the algorithm can work with it.

n.episodes
[integer(1)] Number of episodes.

discount
[numeric(1) in [0, 1]] Discount factor.

alpha
[numeric(1)] Learning rate (step size) for the policy.

beta
[numeric(1)] Learning rate (step size) for the critic.

lambda
[numeric(1) in [0, 1]] Eligibility trace decay parameter.

Value

[list] Returns a list with policy and value function parameters and some statistics about learning behaviour, e.g. the number of steps and return per episode.

Details

When using a gaussian policy mean and variance will be parametrized as a linear combination of features mu = x * w_1 and sigma = exp(x * w_2).

Eligibility traces are used for both policy parameters and value function parameters.

References

Sutton and Barto (Book draft 2017): Reinforcement Learning: An Introduction. Chapter 13

Examples

# Variant of cliff walking
library(reinforcelearn)
rewardFun = function(state, action, n.state) {
  if (n.state %in% 37:46) {
    return(- 100)
  } else {
    return(- 1)
  }
}
env = makeGridworld(shape = c(4, 12), goal.states = 47,
  cliff.states = 37:46, reward.step = - 1, reward.cliff = - 100,
  cliff.transition.done = TRUE, initial.state = 36, sampleReward = rewardFun)

res = actorCritic(env, n.episodes = 100)

#> Episode 1 finished after 1 steps.

#> Episode 2 finished after 21 steps.

#> Episode 3 finished after 23 steps.

#> Episode 4 finished after 1 steps.

#> Episode 5 finished after 9 steps.

#> Episode 6 finished after 66 steps.

#> Episode 7 finished after 94 steps.

#> Episode 8 finished after 56 steps.

#> Episode 9 finished after 6 steps.

#> Episode 10 finished after 34 steps.

#> Episode 11 finished after 45 steps.

#> Episode 12 finished after 37 steps.

#> Episode 13 finished after 107 steps.

#> Episode 14 finished after 65 steps.

#> Episode 15 finished after 81 steps.

#> Episode 16 finished after 53 steps.

#> Episode 17 finished after 159 steps.

#> Episode 18 finished after 115 steps.

#> Episode 19 finished after 54 steps.

#> Episode 20 finished after 164 steps.

#> Episode 21 finished after 73 steps.

#> Episode 22 finished after 302 steps.

#> Episode 23 finished after 149 steps.

#> Episode 24 finished after 105 steps.

#> Episode 25 finished after 179 steps.

#> Episode 26 finished after 58 steps.

#> Episode 27 finished after 96 steps.

#> Episode 28 finished after 337 steps.

#> Episode 29 finished after 126 steps.

#> Episode 30 finished after 130 steps.

#> Episode 31 finished after 100 steps.

#> Episode 32 finished after 237 steps.

#> Episode 33 finished after 67 steps.

#> Episode 34 finished after 145 steps.

#> Episode 35 finished after 135 steps.

#> Episode 36 finished after 50 steps.

#> Episode 37 finished after 76 steps.

#> Episode 38 finished after 75 steps.

#> Episode 39 finished after 104 steps.

#> Episode 40 finished after 62 steps.

#> Episode 41 finished after 138 steps.

#> Episode 42 finished after 208 steps.

#> Episode 43 finished after 17 steps.

#> Episode 44 finished after 47 steps.

#> Episode 45 finished after 93 steps.

#> Episode 46 finished after 49 steps.

#> Episode 47 finished after 49 steps.

#> Episode 48 finished after 56 steps.

#> Episode 49 finished after 88 steps.

#> Episode 50 finished after 72 steps.

#> Episode 51 finished after 107 steps.

#> Episode 52 finished after 68 steps.

#> Episode 53 finished after 57 steps.

#> Episode 54 finished after 20 steps.

#> Episode 55 finished after 66 steps.

#> Episode 56 finished after 59 steps.

#> Episode 57 finished after 41 steps.

#> Episode 58 finished after 51 steps.

#> Episode 59 finished after 36 steps.

#> Episode 60 finished after 83 steps.

#> Episode 61 finished after 58 steps.

#> Episode 62 finished after 4 steps.

#> Episode 63 finished after 64 steps.

#> Episode 64 finished after 44 steps.

#> Episode 65 finished after 26 steps.

#> Episode 66 finished after 43 steps.

#> Episode 67 finished after 41 steps.

#> Episode 68 finished after 53 steps.

#> Episode 69 finished after 55 steps.

#> Episode 70 finished after 85 steps.

#> Episode 71 finished after 47 steps.

#> Episode 72 finished after 15 steps.

#> Episode 73 finished after 42 steps.

#> Episode 74 finished after 61 steps.

#> Episode 75 finished after 35 steps.

#> Episode 76 finished after 38 steps.

#> Episode 77 finished after 29 steps.

#> Episode 78 finished after 57 steps.

#> Episode 79 finished after 40 steps.

#> Episode 80 finished after 28 steps.

#> Episode 81 finished after 77 steps.

#> Episode 82 finished after 71 steps.

#> Episode 83 finished after 64 steps.

#> Episode 84 finished after 40 steps.

#> Episode 85 finished after 35 steps.

#> Episode 86 finished after 52 steps.

#> Episode 87 finished after 35 steps.

#> Episode 88 finished after 29 steps.

#> Episode 89 finished after 70 steps.

#> Episode 90 finished after 25 steps.

#> Episode 91 finished after 42 steps.

#> Episode 92 finished after 41 steps.

#> Episode 93 finished after 33 steps.

#> Episode 94 finished after 40 steps.

#> Episode 95 finished after 32 steps.

#> Episode 96 finished after 42 steps.

#> Episode 97 finished after 51 steps.

#> Episode 98 finished after 44 steps.

#> Episode 99 finished after 26 steps.

#> Episode 100 finished after 33 steps.

#----------------
# Mountain Car
m = MountainCar()

# Define preprocessing function (we use grid tiling)
n.tilings = 8
max.size = 4096
iht = IHT(max.size)

position.max = m$state.space.bounds[[1]][2]
position.min = m$state.space.bounds[[1]][1]
velocity.max = m$state.space.bounds[[2]][2]
velocity.min = m$state.space.bounds[[2]][1]
position.scale = n.tilings / (position.max - position.min)
velocity.scale = n.tilings / (velocity.max - velocity.min)

# Scale state first, then get active tiles and return n hot vector
gridTiling = function(state) {
  state = c(position.scale * state[1], velocity.scale * state[2])
  active.tiles = tiles(iht, 8, state)
  makeNHot(active.tiles, max.size, out = "vector")
}

# Linear function approximation and softmax policy
res = actorCritic(m, fun.approx = "linear",
  preprocessState = gridTiling, n.episodes = 50)

#> Episode 1 finished after 1881 steps.

#> Episode 2 finished after 580 steps.

#> Episode 3 finished after 418 steps.

#> Episode 4 finished after 1782 steps.

#> Episode 5 finished after 232 steps.

#> Episode 6 finished after 223 steps.

#> Episode 7 finished after 237 steps.

#> Episode 8 finished after 154 steps.

#> Episode 9 finished after 144 steps.

#> Episode 10 finished after 149 steps.

#> Episode 11 finished after 152 steps.

#> Episode 12 finished after 149 steps.

#> Episode 13 finished after 151 steps.

#> Episode 14 finished after 149 steps.

#> Episode 15 finished after 149 steps.

#> Episode 16 finished after 151 steps.

#> Episode 17 finished after 148 steps.

#> Episode 18 finished after 147 steps.

#> Episode 19 finished after 156 steps.

#> Episode 20 finished after 138 steps.

#> Episode 21 finished after 145 steps.

#> Episode 22 finished after 144 steps.

#> Episode 23 finished after 154 steps.

#> Episode 24 finished after 150 steps.

#> Episode 25 finished after 154 steps.

#> Episode 26 finished after 146 steps.

#> Episode 27 finished after 191 steps.

#> Episode 28 finished after 150 steps.

#> Episode 29 finished after 139 steps.

#> Episode 30 finished after 143 steps.

#> Episode 31 finished after 150 steps.

#> Episode 32 finished after 153 steps.

#> Episode 33 finished after 146 steps.

#> Episode 34 finished after 145 steps.

#> Episode 35 finished after 152 steps.

#> Episode 36 finished after 141 steps.

#> Episode 37 finished after 155 steps.

#> Episode 38 finished after 145 steps.

#> Episode 39 finished after 144 steps.

#> Episode 40 finished after 143 steps.

#> Episode 41 finished after 151 steps.

#> Episode 42 finished after 147 steps.

#> Episode 43 finished after 152 steps.

#> Episode 44 finished after 148 steps.

#> Episode 45 finished after 153 steps.

#> Episode 46 finished after 146 steps.

#> Episode 47 finished after 167 steps.

#> Episode 48 finished after 156 steps.

#> Episode 49 finished after 154 steps.

#> Episode 50 finished after 143 steps.

#----------------
# Mountain Car with continuous action space
m2 = MountainCar(action.space = "Continuous")

# Linear function approximation and gaussian policy
set.seed(123)
res = actorCritic(m, fun.approx = "linear", policy = "gaussian",
  preprocessState = gridTiling, n.episodes = 50)

#> Episode 1 finished after 1219 steps.

#> Episode 2 finished after 87 steps.

#> Episode 3 finished after 78 steps.

#> Episode 4 finished after 402 steps.

#> Episode 5 finished after 103 steps.

#> Episode 6 finished after 133 steps.

#> Episode 7 finished after 206 steps.

#> Episode 8 finished after 64 steps.

#> Episode 9 finished after 142 steps.

#> Episode 10 finished after 215 steps.

#> Episode 11 finished after 132 steps.

#> Episode 12 finished after 127 steps.

#> Episode 13 finished after 129 steps.

#> Episode 14 finished after 122 steps.

#> Episode 15 finished after 123 steps.

#> Episode 16 finished after 114 steps.

#> Episode 17 finished after 114 steps.

#> Episode 18 finished after 102 steps.

#> Episode 19 finished after 106 steps.

#> Episode 20 finished after 108 steps.

#> Episode 21 finished after 104 steps.

#> Episode 22 finished after 126 steps.

#> Episode 23 finished after 63 steps.

#> Episode 24 finished after 147 steps.

#> Episode 25 finished after 106 steps.

#> Episode 26 finished after 64 steps.

#> Episode 27 finished after 115 steps.

#> Episode 28 finished after 94 steps.

#> Episode 29 finished after 67 steps.

#> Episode 30 finished after 86 steps.

#> Episode 31 finished after 52 steps.

#> Episode 32 finished after 90 steps.

#> Episode 33 finished after 49 steps.

#> Episode 34 finished after 49 steps.

#> Episode 35 finished after 44 steps.

#> Episode 36 finished after 44 steps.

#> Episode 37 finished after 45 steps.

#> Episode 38 finished after 44 steps.

#> Episode 39 finished after 46 steps.

#> Episode 40 finished after 43 steps.

#> Episode 41 finished after 48 steps.

#> Episode 42 finished after 46 steps.

#> Episode 43 finished after 49 steps.

#> Episode 44 finished after 50 steps.

#> Episode 45 finished after 43 steps.

#> Episode 46 finished after 47 steps.

#> Episode 47 finished after 42 steps.

#> Episode 48 finished after 48 steps.

#> Episode 49 finished after 47 steps.

#> Episode 50 finished after 50 steps.