Make Reinforcement Learning Environment

This function creates an environment for reinforcement learning. You could either use an environment from OpenAI Gym or specify the transition and reward array of a Markov Decision Process.

makeEnvironment(gym = NULL, transitions = NULL, rewards = NULL,
  initial.state = NULL, reset = NULL, sampleReward = NULL,
  render = TRUE)

Arguments

gym
[character(1)] Name of gym environment, e.g. "CartPole-v0"

transitions
[array (n.states x n.states x n.actions)] Transition array of Markov Decision Process. For each action specifying the probabilities for transitions between states.

rewards
[matrix (n.states x n.actions) or array (n.states x n.states x n.actions)] Reward array: This can be a matrix (n.states x n.actions) or a 3-dimensional array (n.states x n.states x n.actions). The reward will be sampled from the specified array depending on state, action and possibly also the next state.

initial.state
[integer] The starting state if reset is NULL else this argument is unused. If a vector is given a starting state will be randomly sampled from this vector whenever reset is called. Note that states are numerated starting with 0. If initial.state = NULL all non-terminal states are possible initial states.

reset
[function] Function that returns an initial state observation, it takes no arguments.

sampleReward
[function] Function that returns the next reward given the current state, action and next state. Otherwise the reward will be sampled from the reward array of the MDP specified by rewards.

render
[logical(1)] Whether to render the OpenAI Gym environment. If TRUE a python window with a graphical interface opens whenever the step method is called.

Value

[R6 class] Reinforcement Learning Environment.

Details

This function returns an R6 class with a reset and step method. Everytime when you start an episode call the reset method first. Use the step method to interact with the environment. Note that the methods do not return anything, but change the attributes of the R6 class, most importantly the state, reward and done attribute.

Note that all states and actions are numerated starting with 0!

For OpenAI gym environments have a look at https://gym.openai.com/envs and at the description of what requirements to install at https://github.com/openai/gym-http-api.

To use a gym environment please start the python server (gym_http_server.py) in a terminal before calling makeEnvironment. You will find the file in your gym installation or also a copy in the `reinforcelearn` package library. You can also start the server using R code (see examples below).

For a detailed explanation and more examples have a look at the vignette "How to create an environment?".

Methods

envir$step(action)

Takes a step in the environment. Given an action the method returns the next state, reward and if the episode has finished. If a transition array and reward matrix are given, the next step will be sampled from the MDP, else env_step will be called.

envir$reset()

Resets the done flag of the environment and returns an initial state. Useful when starting a new episode.

envir$close()

Close the python window for an OpenAI Gym environment.

See also

Create gridworlds with gridworld. For the mountain car environment have a look at mountainCar.

Examples

# Create a Markov Decision Process.
P = array(0, c(2, 2, 2))
P[, , 1] = matrix(c(0.5, 0.5, 0, 1), 2, 2, byrow = TRUE)
P[, , 2] = matrix(c(0, 1, 0, 1), 2, 2, byrow = TRUE)
R = matrix(c(5, 10, -1, 2), 2, 2, byrow = TRUE)
env = makeEnvironment(transitions = P, rewards = R)

# Specify a custom probability distribution for the starting state.
reset = function() {
  p = c(0.2, 0.8)
  sample(0:1, prob = p, size = 1)
}
env = makeEnvironment(transitions = P, rewards = R, reset = reset)
env$reset()

#> Warning: The starting state is a terminal state!

print(env$state)

#> [1] 1

# Specify a custom reward function.
sampleReward = function(state, action, n.state) {
  if (state == 2 && action == 1L) {
    rexp(1)
  } else {
    rnorm(1)
  }
}
env = makeEnvironment(transitions = P, rewards = R, sampleReward = sampleReward)
env$reset()
env$step(1)
print(env$reward)

#> [1] -0.2713959

# Gridworld Environment
grid = gridworld(shape = c(4, 4), goal.states = 15, initial.state = 0)
# NOT RUN {
# Create an OpenAI Gym environment.
# Make sure you have Python and Gym installed.
# Start server from within R.
package.path = system.file(package = "reinforcelearn")
path2pythonfile = paste0(package.path, "/gym_http_server.py")
system2("python", args = path2pythonfile, stdout = NULL,
  wait = FALSE, invisible = FALSE)

env = makeEnvironment("CartPole-v0", render = FALSE)
env$reset()
print(env)

# Create the MountainCar environment which has a continuous state space.
env = makeEnvironment("MountainCar-v0")
env$state.space
env$state.space.bounds

# Take random actions for 200 steps.
env$reset()
for (i in 1:200) {
  action = sample(env$actions, 1)
  env$step(action)
}
env$close()
# }