Q(sigma) / Q-Learning / Sarsa / Expected Sarsa

Value-based reinforcement learning control algorithms.

qSigma(envir, fun.approx = "table", preprocessState = identity,
  model = NULL, initial.value = NULL, n.states = NULL, n.episodes = 100,
  sigma = 1, target.policy = "egreedy", lambda = 0,
  eligibility.type = 0, learning.rate = 0.1, epsilon = 0.1,
  discount = 1, double.learning = FALSE, replay.memory = NULL,
  replay.memory.size = 1, batch.size = 1, alpha = 0, theta = 0.01,
  updateEpsilon = identity2, updateSigma = identity2,
  updateLambda = identity2, updateAlpha = identity2,
  updateLearningRate = identity2)

qlearning(envir, fun.approx = "table", preprocessState = identity,
  model = NULL, initial.value = NULL, n.states = NULL, n.episodes = 100,
  lambda = 0, eligibility.type = 0, learning.rate = 0.1, epsilon = 0.1,
  discount = 1, double.learning = FALSE, replay.memory = NULL,
  replay.memory.size = 1, batch.size = 1, alpha = 0, theta = 0.01,
  updateEpsilon = identity2, updateSigma = identity2,
  updateLambda = identity2, updateAlpha = identity2,
  updateLearningRate = identity2)

sarsa(envir, fun.approx = "table", preprocessState = identity,
  model = NULL, initial.value = NULL, n.states = NULL, n.episodes = 100,
  lambda = 0, eligibility.type = 0, learning.rate = 0.1, epsilon = 0.1,
  discount = 1, double.learning = FALSE, replay.memory = NULL,
  replay.memory.size = 1, batch.size = 1, alpha = 0, theta = 0.01,
  updateEpsilon = identity2, updateSigma = identity2,
  updateLambda = identity2, updateAlpha = identity2,
  updateLearningRate = identity2)

expectedSarsa(envir, fun.approx = "table", preprocessState = identity,
  model = NULL, initial.value = NULL, n.states = NULL, n.episodes = 100,
  target.policy = "egreedy", lambda = 0, eligibility.type = 0,
  learning.rate = 0.1, epsilon = 0.1, discount = 1,
  double.learning = FALSE, replay.memory = NULL, replay.memory.size = 1,
  batch.size = 1, alpha = 0, theta = 0.01, updateEpsilon = identity2,
  updateSigma = identity2, updateLambda = identity2,
  updateAlpha = identity2, updateLearningRate = identity2)

Arguments

envir
[R6 class] The reinforcement learning environment created by makeEnvironment.

fun.approx
[character(1)] How to represent the value function? Currently "table" and "linear" are supported.

preprocessState
[function] A function that takes the state observation returned from the environment as an input and preprocesses this in a way the algorithm can work with it.

model
[any] Currently unused parameter.

initial.value
[numeric] Initial value function matrix or weight matrix. If NULL weights will be initialized to 0. Only used for tabular or linear function approximation.

n.states
[integer(1)] Number of states for tabular value function. Only needed if the state space is continuous, but will be transformed by preprocessState, so that a single integer value is returned. Else the number of states is accessed from the envir$n.states argument.

n.episodes
[integer(1)] Number of episodes.

sigma
[numeric(1) in [0, 1]] Parameter of the Q(sigma) algorithm. It controls if the temporal-difference target is equal to the sarsa target (for sigma = 1) or the expected sarsa target (for sigma = 0). For intermediate values of sigma a weighted mean between the two targets is used.

target.policy
[character(1)] Should the expected sarsa target be computed on-policy using the epsilon-greedy behavior policy (target.policy = "egreedy") or off-policy using a greedy target policy (target.policy = "greedy")?

lambda
[numeric(1) in [0, 1]] Eligibility trace decay parameter.

eligibility.type
[numeric(1)] Type of eligibility trace, use eligibility.type = 1 for replacing traces, eligibility.type = 0 for accumulating traces or intermediate values for a mixture between both.

learning.rate
[numeric(1)] Learning rate used for gradient descent.

epsilon
[numeric(1) in [0, 1]] Ratio of random exploration in epsilon-greedy action selection.

discount
[numeric(1) in [0, 1]] Discount factor.

double.learning
[logical(1)] Should double learning be used?

replay.memory
[list] Initial replay memory, which can be passed on. Each list element must be a list containing state, action, reward and next.state.

replay.memory.size
[integer(1)] Size of the replay memory. Only used if replay.memory is NULL. Then the replay memory will be initially filled with experiences following a uniform random policy.

batch.size
[integer(1)] Batch size, how many experiences are sampled from the replay memory at each step? Must be smaller than the size of the replay memory!

alpha
[numeric(1) in [0, 1]] If alpha = 0 sampling from the replay memory will be uniform, otherwise observations with a high error will be proportionally prioritized.

theta
[numeric(1) in (0, 1]] theta is a small positive constant that prevents the edge-case of transitions in the replay memory not being revisited once their error is zero.

updateEpsilon
[function] A function that updates epsilon. It takes two arguments, epsilon and the current number of episodes which are finished, and returns the new epsilon value.

updateSigma
[function] A function that updates sigma. It takes two arguments, sigma and the current number of episodes which are finished, and returns the new sigma value.

updateLambda
[function] A function that updates lambda. It takes two arguments, lambda and the current number of episodes which are finished, and returns the new lambda value.

updateAlpha
[function] A function that updates alpha. It takes two arguments, alpha and the current number of episodes which are finished, and returns the new alpha value.

updateLearningRate
[function] A function that updates the learning rate. It takes two arguments, learning.rate and the current number of episodes which are finished, and returns the new learning.rate value.

Value

[list(4)] Returns the action value function or model parameters [matrix] and the return and number of steps per episode [numeric]. For Double Learning both value functions will be returned.

Details

Q(sigma), Q-Learning, Expected Sarsa and Sarsa build a family of temporal-difference learning algorithms, which can be used to find the optimal action value function using the principle of generalized policy iteration. Q(sigma) subsumes Q-Learning, Sarsa and Expected Sarsa as special cases. For more details on the algorithm read De Asis et al. (2017). Expected Sarsa can be used on-policy if target.policy == "egreedy" else it is identical to Q-Learning.

When fun.approx == "table" the action value function will be represented as a matrix, but you can also use a linear combination of features for function approximation.

The raw state observation returned from the environment must be preprocessed using the preprocessState argument. This function takes the state observation as input and returns a preprocessed state which can be directly used by the function approximator. To use a tabular value function preprocessState must return an integer value between [0, number of states - 1]. For linear function approximation the output of preprocessState must be a matrix with one row.

Experience replay can be used by specifying a prefilled replay memory using the replay.memory argument or by specifying the length of the replay memory, which is then filled with experience using a random uniform policy. Sampling can also be prioritized by the error as proposed by Schaul et al. (2016).

Note that eligibility traces cannot be used with experience replay.

The hyperparameters epsilon, sigma, alpha, lambda and learning.rate can be changed over time. Therefore pass on functions that return a new value of the hyperparameter. These updates will be applied after each episode. First argument of the function must be the parameter itself, the second argument the current episode number.

References

De Asis et al. (2017): Multi-step Reinforcement Learning: A Unifying Algorithm

Hasselt et al. (2010): Double Q-Learning

Mnih et al. (2013): Playing Atari with Deep Reinforcement Learning

Schaul et al. (2016): Prioritized Experience Replay

Sutton and Barto (Book draft 2017): Reinforcement Learning: An Introduction

Examples

# Solve Windy Gridworld
env = makeEnvironment("WindyGridworld")

res = qlearning(env, n.episodes = 20)

#> Episode 1 finished after 854 steps with a return of -854

#> Episode 2 finished after 1236 steps with a return of -1236

#> Episode 3 finished after 1063 steps with a return of -1063

#> Episode 4 finished after 367 steps with a return of -367

#> Episode 5 finished after 427 steps with a return of -427

#> Episode 6 finished after 357 steps with a return of -357

#> Episode 7 finished after 103 steps with a return of -103

#> Episode 8 finished after 424 steps with a return of -424

#> Episode 9 finished after 261 steps with a return of -261

#> Episode 10 finished after 251 steps with a return of -251

#> Episode 11 finished after 230 steps with a return of -230

#> Episode 12 finished after 226 steps with a return of -226

#> Episode 13 finished after 141 steps with a return of -141

#> Episode 14 finished after 112 steps with a return of -112

#> Episode 15 finished after 169 steps with a return of -169

#> Episode 16 finished after 161 steps with a return of -161

#> Episode 17 finished after 287 steps with a return of -287

#> Episode 18 finished after 106 steps with a return of -106

#> Episode 19 finished after 79 steps with a return of -79

#> Episode 20 finished after 35 steps with a return of -35

print(res$steps)

#>  [1]  854 1236 1063  367  427  357  103  424  261  251  230  226  141  112  169
#> [16]  161  287  106   79   35

# State value function
print(matrix(round(apply(res$Q1, 1, max), 1), ncol = 10, byrow = TRUE))

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,] -4.1 -4.3 -4.8 -5.6 -6.3 -6.7 -6.8 -6.3 -5.6  -4.8
#> [2,] -3.9 -4.0 -4.1 -4.4 -4.1 -3.2 -1.9 -3.4 -4.0  -3.9
#> [3,] -3.5 -3.5 -3.4 -3.3 -2.6 -1.1 -0.5 -1.7 -3.1  -3.1
#> [4,] -3.3 -3.1 -2.8 -2.4 -1.3 -0.3 -0.1  0.0 -1.9  -2.3
#> [5,] -2.8 -2.6 -2.2 -1.7 -0.7 -0.2  0.0 -0.2 -0.8  -1.6
#> [6,] -2.4 -2.1 -1.7 -1.2 -0.4  0.0  0.0  0.0 -0.5  -0.9
#> [7,] -2.1 -1.9 -1.4 -0.7  0.0  0.0  0.0  0.0 -0.1  -0.5

# Policy
policy = max.col(res$Q1) - 1L
print(matrix(policy, ncol = 10, byrow = TRUE))

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    0    1    0    0    3    0    3    2    1     3
#> [2,]    3    0    2    0    3    3    3    3    3     3
#> [3,]    3    0    2    0    0    3    1    1    0     3
#> [4,]    3    3    2    0    1    0    3    1    3     3
#> [5,]    3    0    0    1    1    2    1    3    0     3
#> [6,]    3    1    3    1    0    0    0    3    2     0
#> [7,]    1    1    3    2    1    1    2    1    2     3

# Decay epsilon over time. Each 10 episodes epsilon will be halfed.
decayEpsilon = function(epsilon, i) {
  if (i %% 10 == 0) {
    epsilon = epsilon * 0.5
  }
  epsilon
}

res = expectedSarsa(env, epsilon = 0.5, updateEpsilon = decayEpsilon, n.episodes = 20)

#> Episode 1 finished after 3351 steps with a return of -3351

#> Episode 2 finished after 111 steps with a return of -111

#> Episode 3 finished after 77 steps with a return of -77

#> Episode 4 finished after 581 steps with a return of -581

#> Episode 5 finished after 1746 steps with a return of -1746

#> Episode 6 finished after 203 steps with a return of -203

#> Episode 7 finished after 59 steps with a return of -59

#> Episode 8 finished after 153 steps with a return of -153

#> Episode 9 finished after 298 steps with a return of -298

#> Episode 10 finished after 387 steps with a return of -387

#> Episode 11 finished after 23 steps with a return of -23

#> Episode 12 finished after 294 steps with a return of -294

#> Episode 13 finished after 112 steps with a return of -112

#> Episode 14 finished after 185 steps with a return of -185

#> Episode 15 finished after 205 steps with a return of -205

#> Episode 16 finished after 114 steps with a return of -114

#> Episode 17 finished after 88 steps with a return of -88

#> Episode 18 finished after 167 steps with a return of -167

#> Episode 19 finished after 21 steps with a return of -21

#> Episode 20 finished after 76 steps with a return of -76

# Solve the Mountain Car problem using linear function approximation
m = makeEnvironment("MountainCar")

# Define preprocessing function (we use grid tiling)
n.tilings = 8
max.size = 4096
iht = IHT(max.size)

position.max = m$state.space.bounds[[1]][2]
position.min = m$state.space.bounds[[1]][1]
velocity.max = m$state.space.bounds[[2]][2]
velocity.min = m$state.space.bounds[[2]][1]
position.scale = n.tilings / (position.max - position.min)
velocity.scale = n.tilings / (velocity.max - velocity.min)

# Scale state first, then get active tiles and return n hot vector
gridTiling = function(state) {
  state = c(position.scale * state[1], velocity.scale * state[2])
  active.tiles = tiles(iht, 8, state)
  nHot(active.tiles, max.size, out = "vector")
}

set.seed(123)
res = sarsa(m, fun.approx = "linear", preprocessState = gridTiling, n.episodes = 20)

#> Episode 1 finished after 1036 steps with a return of -1035

#> Episode 2 finished after 712 steps with a return of -711

#> Episode 3 finished after 564 steps with a return of -563

#> Episode 4 finished after 394 steps with a return of -393

#> Episode 5 finished after 485 steps with a return of -484

#> Episode 6 finished after 322 steps with a return of -321

#> Episode 7 finished after 427 steps with a return of -426

#> Episode 8 finished after 321 steps with a return of -320

#> Episode 9 finished after 232 steps with a return of -231

#> Episode 10 finished after 288 steps with a return of -287

#> Episode 11 finished after 192 steps with a return of -191

#> Episode 12 finished after 236 steps with a return of -235

#> Episode 13 finished after 302 steps with a return of -301

#> Episode 14 finished after 234 steps with a return of -233

#> Episode 15 finished after 198 steps with a return of -197

#> Episode 16 finished after 198 steps with a return of -197

#> Episode 17 finished after 249 steps with a return of -248

#> Episode 18 finished after 317 steps with a return of -316

#> Episode 19 finished after 258 steps with a return of -257

#> Episode 20 finished after 241 steps with a return of -240

print(res$returns)

#>  [1] -1035  -711  -563  -393  -484  -321  -426  -320  -231  -287  -191  -235
#> [13]  -301  -233  -197  -197  -248  -316  -257  -240