Q(sigma) / Q-Learning / Sarsa / Expected Sarsa

Value-based reinforcement learning control algorithms.

qSigma(envir, fun.approx = "table", preprocessState = identity,
  model = NULL, initial.value = NULL, n.states = NULL, n.episodes = 100,
  sigma = 1, target.policy = "egreedy", lambda = 0,
  eligibility.type = 0, learning.rate = 0.1, epsilon = 0.1,
  discount = 1, double.learning = FALSE, replay.memory = NULL,
  replay.memory.size = 1, batch.size = 1, alpha = 0, theta = 0.01,
  updateEpsilon = identity2, updateSigma = identity2,
  updateLambda = identity2, updateAlpha = identity2,
  updateLearningRate = identity2)

qlearning(envir, fun.approx = "table", preprocessState = identity,
  model = NULL, initial.value = NULL, n.states = NULL, n.episodes = 100,
  lambda = 0, eligibility.type = 0, learning.rate = 0.1, epsilon = 0.1,
  discount = 1, double.learning = FALSE, replay.memory = NULL,
  replay.memory.size = 1, batch.size = 1, alpha = 0, theta = 0.01,
  updateEpsilon = identity2, updateSigma = identity2,
  updateLambda = identity2, updateAlpha = identity2,
  updateLearningRate = identity2)

sarsa(envir, fun.approx = "table", preprocessState = identity,
  model = NULL, initial.value = NULL, n.states = NULL, n.episodes = 100,
  lambda = 0, eligibility.type = 0, learning.rate = 0.1, epsilon = 0.1,
  discount = 1, double.learning = FALSE, replay.memory = NULL,
  replay.memory.size = 1, batch.size = 1, alpha = 0, theta = 0.01,
  updateEpsilon = identity2, updateSigma = identity2,
  updateLambda = identity2, updateAlpha = identity2,
  updateLearningRate = identity2)

expectedSarsa(envir, fun.approx = "table", preprocessState = identity,
  model = NULL, initial.value = NULL, n.states = NULL, n.episodes = 100,
  target.policy = "egreedy", lambda = 0, eligibility.type = 0,
  learning.rate = 0.1, epsilon = 0.1, discount = 1,
  double.learning = FALSE, replay.memory = NULL, replay.memory.size = 1,
  batch.size = 1, alpha = 0, theta = 0.01, updateEpsilon = identity2,
  updateSigma = identity2, updateLambda = identity2,
  updateAlpha = identity2, updateLearningRate = identity2)

Arguments

envir
[R6 class] The reinforcement learning environment created by makeEnvironment.

fun.approx
[character(1)] How to represent the value function? Currently "table", "linear" and "neural.network" are supported.

preprocessState
[function] A function that takes the state observation returned from the environment as an input and preprocesses this in a way the algorithm can work with it.

model
[keras model] A neural network model specified using the keras package.

initial.value
[numeric] Initial value function matrix or weight matrix. If NULL weights will be initialized to 0. Only used for tabular or linear function approximation.

n.states
[integer(1)] Number of states for tabular value function. Only needed if the state space is continuous, but will be transformed by preprocessState, so that a single integer value is returned. Else the number of states is accessed from the envir$n.states argument.

n.episodes
[integer(1)] Number of episodes.

sigma
[numeric(1) in [0, 1]] Parameter of the Q(sigma) algorithm. It controls if the temporal-difference target is equal to the sarsa target (for sigma = 1) or the expected sarsa target (for sigma = 0). For intermediate values of sigma a weighted mean between the two targets is used.

target.policy
[character(1)] Should the expected sarsa target be computed on-policy using the epsilon-greedy behavior policy (target.policy = "egreedy") or off-policy using a greedy target policy (target.policy = "greedy")?

lambda
[numeric(1) in [0, 1]] Eligibility trace decay parameter.

eligibility.type
[numeric(1)] Type of eligibility trace, use eligibility.type = 1 for replacing traces, eligibility.type = 0 for accumulating traces or intermediate values for a mixture between both.

learning.rate
[numeric(1)] Learning rate used for gradient descent.

epsilon
[numeric(1) in [0, 1]] Ratio of random exploration in epsilon-greedy action selection.

discount
[numeric(1) in [0, 1]] Discount factor.

double.learning
[logical(1)] Should double learning be used?

replay.memory
[list] Initial replay memory, which can be passed on. Each list element must be a list containing state, action, reward and next.state.

replay.memory.size
[integer(1)] Size of the replay memory. Only used if replay.memory is NULL. Then the replay memory will be initially filled with experiences following a uniform random policy.

batch.size
[integer(1)] Batch size, how many experiences are sampled from the replay memory at each step? Must be smaller than the size of the replay memory!

alpha
[numeric(1) in [0, 1]] If alpha = 0 sampling from the replay memory will be uniform, otherwise observations with a high error will be proportionally prioritized.

theta
[numeric(1) in (0, 1]] theta is a small positive constant that prevents the edge-case of transitions in the replay memory not being revisited once their error is zero.

updateEpsilon
[function] A function that updates epsilon. It takes two arguments, epsilon and the current number of episodes which are finished, and returns the new epsilon value.

updateSigma
[function] A function that updates sigma. It takes two arguments, sigma and the current number of episodes which are finished, and returns the new sigma value.

updateLambda
[function] A function that updates lambda. It takes two arguments, lambda and the current number of episodes which are finished, and returns the new lambda value.

updateAlpha
[function] A function that updates alpha. It takes two arguments, alpha and the current number of episodes which are finished, and returns the new alpha value.

updateLearningRate
[function] A function that updates the learning rate. It takes two arguments, learning.rate and the current number of episodes which are finished, and returns the new learning.rate value.

Value

[list(4)] Returns the action value function or model parameters [matrix] and the return and number of steps per episode [numeric]. For Double Learning both value functions will be returned.

Details

Q(sigma), Q-Learning, Expected Sarsa and Sarsa build a family of temporal-difference learning algorithms, which can be used to find the optimal action value function using the principle of generalized policy iteration. Q(sigma) subsumes Q-Learning, Sarsa and Expected Sarsa as special cases. For more details on the algorithm read De Asis et al. (2017). Expected Sarsa can be used on-policy if target.policy == "egreedy" else it is identical to Q-Learning.

When fun.approx == "table" the action value function will be represented as a matrix, but you can also use a linear combination of features or a neural network for function approximation. For a neural network you need to pass on a keras model via the model argument.

The raw state observation returned from the environment must be preprocessed using the preprocessState argument. This function takes the state observation as input and returns a preprocessed state which can be directly used by the function approximator. To use a tabular value function preprocessState must return an integer value between [0, number of states - 1]. For linear function approximation the output of preprocessState must be a matrix with one row. For a neural network you have to make sure that the dimensions of the preprocessed state and the neural network match, so that model$predict(preprocessState(envir$state)) works.

Experience replay can be used by specifying a prefilled replay memory using the replay.memory argument or by specifying the length of the replay memory, which is then filled with experience using a random uniform policy. Sampling can also be prioritized by the error as proposed by Schaul et al. (2016).

Note that eligibility traces cannot be used with experience replay.

The hyperparameters epsilon, sigma, alpha, lambda and learning.rate can be changed over time. Therefore pass on functions that return a new value of the hyperparameter. These updates will be applied after each episode. First argument of the function must be the parameter itself, the second argument the current episode number.

For neural networks the options experience replay, eligibility traces and double learning are currently not available.

References

De Asis et al. (2017): Multi-step Reinforcement Learning: A Unifying Algorithm

Hasselt et al. (2010): Double Q-Learning

Mnih et al. (2013): Playing Atari with Deep Reinforcement Learning

Schaul et al. (2016): Prioritized Experience Replay

Sutton and Barto (Book draft 2017): Reinforcement Learning: An Introduction

Examples

# Solve Windy Gridworld
env = windyGridworld()

res = qlearning(env, n.episodes = 20)

#> Episode 1 finished after 794 steps with a return of -794

#> Episode 2 finished after 1701 steps with a return of -1701

#> Episode 3 finished after 746 steps with a return of -746

#> Episode 4 finished after 468 steps with a return of -468

#> Episode 5 finished after 319 steps with a return of -319

#> Episode 6 finished after 352 steps with a return of -352

#> Episode 7 finished after 366 steps with a return of -366

#> Episode 8 finished after 256 steps with a return of -256

#> Episode 9 finished after 131 steps with a return of -131

#> Episode 10 finished after 136 steps with a return of -136

#> Episode 11 finished after 172 steps with a return of -172

#> Episode 12 finished after 232 steps with a return of -232

#> Episode 13 finished after 219 steps with a return of -219

#> Episode 14 finished after 202 steps with a return of -202

#> Episode 15 finished after 40 steps with a return of -40

#> Episode 16 finished after 85 steps with a return of -85

#> Episode 17 finished after 158 steps with a return of -158

#> Episode 18 finished after 200 steps with a return of -200

#> Episode 19 finished after 120 steps with a return of -120

#> Episode 20 finished after 140 steps with a return of -140

print(res$steps)

#>  [1]  794 1701  746  468  319  352  366  256  131  136  172  232  219  202   40
#> [16]   85  158  200  120  140

# State value function
print(matrix(round(apply(res$Q1, 1, max), 1), ncol = 10, byrow = TRUE))

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,] -4.2 -4.4 -4.8 -5.6 -6.3 -6.7 -6.7 -6.2 -5.5  -4.7
#> [2,] -3.9 -4.0 -4.1 -4.4 -4.0 -3.4 -1.8 -3.4 -3.8  -3.8
#> [3,] -3.6 -3.5 -3.4 -3.3 -2.6 -1.3 -0.3 -1.8 -2.7  -3.0
#> [4,] -3.3 -3.1 -2.8 -2.5 -1.4 -0.4  0.0  0.0 -1.8  -2.2
#> [5,] -2.8 -2.6 -2.2 -1.7 -0.7 -0.2  0.0 -0.2 -0.8  -1.5
#> [6,] -2.5 -2.2 -1.7 -1.2 -0.4  0.0  0.0  0.0 -0.5  -0.9
#> [7,] -2.2 -1.9 -1.4 -0.8  0.0  0.0  0.0  0.0 -0.1  -0.5

# Policy
policy = max.col(res$Q1) - 1L
print(matrix(policy, ncol = 10, byrow = TRUE))

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    3    0    3    1    1    0    0    1    1     3
#> [2,]    3    0    3    3    3    0    2    1    0     3
#> [3,]    2    1    3    3    2    1    0    3    3     3
#> [4,]    3    3    1    0    1    1    3    2    2     2
#> [5,]    2    1    3    2    3    2    0    0    0     2
#> [6,]    3    0    0    0    3    3    0    3    3     1
#> [7,]    1    0    1    0    2    2    3    2    2     1

# Decay epsilon over time. Each 10 episodes epsilon will be halfed.
decayEpsilon = function(epsilon, i) {
  if (i %% 10 == 0) {
    epsilon = epsilon * 0.5
  }
  epsilon
}

res = expectedSarsa(env, epsilon = 0.5, updateEpsilon = decayEpsilon, n.episodes = 20)

#> Episode 1 finished after 478 steps with a return of -478

#> Episode 2 finished after 2727 steps with a return of -2727

#> Episode 3 finished after 884 steps with a return of -884

#> Episode 4 finished after 365 steps with a return of -365

#> Episode 5 finished after 530 steps with a return of -530

#> Episode 6 finished after 312 steps with a return of -312

#> Episode 7 finished after 353 steps with a return of -353

#> Episode 8 finished after 210 steps with a return of -210

#> Episode 9 finished after 242 steps with a return of -242

#> Episode 10 finished after 293 steps with a return of -293

#> Episode 11 finished after 100 steps with a return of -100

#> Episode 12 finished after 140 steps with a return of -140

#> Episode 13 finished after 87 steps with a return of -87

#> Episode 14 finished after 146 steps with a return of -146

#> Episode 15 finished after 79 steps with a return of -79

#> Episode 16 finished after 84 steps with a return of -84

#> Episode 17 finished after 156 steps with a return of -156

#> Episode 18 finished after 99 steps with a return of -99

#> Episode 19 finished after 226 steps with a return of -226

#> Episode 20 finished after 74 steps with a return of -74

# Solve the Mountain Car problem using linear function approximation
m = mountainCar()

# Define preprocessing function (we use grid tiling)
n.tilings = 8
max.size = 4096
iht = IHT(max.size)

position.max = m$state.space.bounds[[1]][2]
position.min = m$state.space.bounds[[1]][1]
velocity.max = m$state.space.bounds[[2]][2]
velocity.min = m$state.space.bounds[[2]][1]
position.scale = n.tilings / (position.max - position.min)
velocity.scale = n.tilings / (velocity.max - velocity.min)

# Scale state first, then get active tiles and return n hot vector
gridTiling = function(state) {
  state = c(position.scale * state[1], velocity.scale * state[2])
  active.tiles = tiles(iht, 8, state)
  nHot(active.tiles, max.size, out = "vector")
}

set.seed(123)
res = sarsa(m, fun.approx = "linear", preprocessState = gridTiling, n.episodes = 20)

#> Episode 1 finished after 840 steps with a return of -839

#> Episode 2 finished after 784 steps with a return of -783

#> Episode 3 finished after 684 steps with a return of -683

#> Episode 4 finished after 694 steps with a return of -693

#> Episode 5 finished after 318 steps with a return of -317

#> Episode 6 finished after 475 steps with a return of -474

#> Episode 7 finished after 270 steps with a return of -269

#> Episode 8 finished after 274 steps with a return of -273

#> Episode 9 finished after 235 steps with a return of -234

#> Episode 10 finished after 199 steps with a return of -198

#> Episode 11 finished after 276 steps with a return of -275

#> Episode 12 finished after 160 steps with a return of -159

#> Episode 13 finished after 213 steps with a return of -212

#> Episode 14 finished after 304 steps with a return of -303

#> Episode 15 finished after 216 steps with a return of -215

#> Episode 16 finished after 235 steps with a return of -234

#> Episode 17 finished after 224 steps with a return of -223

#> Episode 18 finished after 180 steps with a return of -179

#> Episode 19 finished after 153 steps with a return of -152

#> Episode 20 finished after 222 steps with a return of -221

print(res$returns)

#>  [1] -839 -783 -683 -693 -317 -474 -269 -273 -234 -198 -275 -159 -212 -303 -215
#> [16] -234 -223 -179 -152 -221

# NOT RUN {
env = gridworld(c(3, 3), goal.states = 8, initial.state = 0)

# Use a neural network as function approximator
oneHot = function(state) {
  one.hot = matrix(rep(0, 9), nrow = 1)
  one.hot[1, state + 1] = 1
  one.hot
}

# Define keras model
library(keras)
model = keras_model_sequential()
model %>% layer_dense(units = 4, activation = 'linear', input_shape = c(9))

res = qSigma(env, fun.approx = "neural.network", model = model,
  preprocessState = oneHot, n.episodes = 20)
# }