TD(lambda)

Estimate the state value function for a given policy using temporal difference learning.

td(envir, policy, fun.approx = "table", preprocessState = identity,
  initial.value = NULL, n.episodes = NULL, n.steps = NULL, discount = 1,
  lambda = 0, eligibility.type = 0, learning.rate = 0.1)

Arguments

envir
[R6 class] The reinforcement learning environment created by makeEnvironment.

policy
[matrix | function] A policy specified as a probability matrix (states x actions) or a function returning an action given a preprocessed state observation (with linear function approximation).

fun.approx
[character(1)] How to represent the value function? Currently "table", "linear" and "neural.network" are supported.

preprocessState
[function] A function that takes the state observation returned from the environment as an input and preprocesses this in a way the algorithm can work with it.

initial.value
[numeric] Initial value function matrix or weight matrix. If NULL weights will be initialized to 0. Only used for tabular or linear function approximation.

n.episodes
[integer(1)] Number of episodes.

n.steps
[integer(1)] Number of evaluations (steps in the environment).

discount
[numeric(1) in [0, 1]] Discount factor.

lambda
[numeric(1) in [0, 1]] Eligibility trace decay parameter.

eligibility.type
[numeric(1)] Type of eligibility trace, use eligibility.type = 1 for replacing traces, eligibility.type = 0 for accumulating traces or intermediate values for a mixture between both.

learning.rate
[numeric(1)] Learning rate used for gradient descent.

Value

[list] Returns the state value function V and some statistics about learning behavior, e.g. the number of steps and return per episode.

Details

Can be used for episodic or continuing environments, please specify either a maximum number of steps or a maximum number of episodes.

References

Sutton and Barto (Book draft 2017): Reinforcement Learning: An Introduction. Chapter 6

Examples

# Random Walk Task (Sutton & Barto Example 6.2)
P = array(dim = c(7, 7, 2))
P[, , 1] = matrix(c(rep(c(1, rep(0, 6)), 2), c(0, 1, rep(0, 5)),
  c(0, 0, 1, rep(0, 4)), c(rep(0, 3), 1, rep(0, 3)), c(rep(0, 4), 1, rep(0, 2)),
  c(rep(0, 6), 1)), ncol = 7, byrow = TRUE)
P[, , 2] = matrix(c(c(1, rep(0, 6)), c(0, 0, 1, rep(0, 4)),
  c(rep(0, 3), 1, rep(0, 3)), c(rep(0, 4), 1, rep(0, 2)),
  c(rep(0, 5), 1, 0), c(rep(0, 6), 1), c(rep(0, 6), 1)), ncol = 7, byrow = TRUE)
R = matrix(c(rep(0, 12), 1, 0), ncol = 2)
env = makeEnvironment(transitions = P, rewards = R, initial.state = 3)

# Uniform random policy
random.policy = matrix(1 / env$n.actions, nrow = env$n.states,
  ncol = env$n.actions)

# Estimate state value function with TD(0)
res = td(env, random.policy, n.episodes = 100)

#> Episode 1 finished after 5 steps.

#> Episode 2 finished after 7 steps.

#> Episode 3 finished after 3 steps.

#> Episode 4 finished after 13 steps.

#> Episode 5 finished after 11 steps.

#> Episode 6 finished after 27 steps.

#> Episode 7 finished after 31 steps.

#> Episode 8 finished after 3 steps.

#> Episode 9 finished after 9 steps.

#> Episode 10 finished after 17 steps.

#> Episode 11 finished after 11 steps.

#> Episode 12 finished after 7 steps.

#> Episode 13 finished after 5 steps.

#> Episode 14 finished after 7 steps.

#> Episode 15 finished after 13 steps.

#> Episode 16 finished after 3 steps.

#> Episode 17 finished after 5 steps.

#> Episode 18 finished after 7 steps.

#> Episode 19 finished after 5 steps.

#> Episode 20 finished after 3 steps.

#> Episode 21 finished after 13 steps.

#> Episode 22 finished after 7 steps.

#> Episode 23 finished after 9 steps.

#> Episode 24 finished after 3 steps.

#> Episode 25 finished after 21 steps.

#> Episode 26 finished after 3 steps.

#> Episode 27 finished after 5 steps.

#> Episode 28 finished after 5 steps.

#> Episode 29 finished after 3 steps.

#> Episode 30 finished after 7 steps.

#> Episode 31 finished after 7 steps.

#> Episode 32 finished after 9 steps.

#> Episode 33 finished after 27 steps.

#> Episode 34 finished after 13 steps.

#> Episode 35 finished after 5 steps.

#> Episode 36 finished after 3 steps.

#> Episode 37 finished after 3 steps.

#> Episode 38 finished after 11 steps.

#> Episode 39 finished after 7 steps.

#> Episode 40 finished after 3 steps.

#> Episode 41 finished after 3 steps.

#> Episode 42 finished after 7 steps.

#> Episode 43 finished after 7 steps.

#> Episode 44 finished after 7 steps.

#> Episode 45 finished after 9 steps.

#> Episode 46 finished after 3 steps.

#> Episode 47 finished after 5 steps.

#> Episode 48 finished after 11 steps.

#> Episode 49 finished after 3 steps.

#> Episode 50 finished after 7 steps.

#> Episode 51 finished after 19 steps.

#> Episode 52 finished after 3 steps.

#> Episode 53 finished after 3 steps.

#> Episode 54 finished after 3 steps.

#> Episode 55 finished after 13 steps.

#> Episode 56 finished after 13 steps.

#> Episode 57 finished after 15 steps.

#> Episode 58 finished after 5 steps.

#> Episode 59 finished after 3 steps.

#> Episode 60 finished after 3 steps.

#> Episode 61 finished after 7 steps.

#> Episode 62 finished after 3 steps.

#> Episode 63 finished after 3 steps.

#> Episode 64 finished after 3 steps.

#> Episode 65 finished after 9 steps.

#> Episode 66 finished after 3 steps.

#> Episode 67 finished after 7 steps.

#> Episode 68 finished after 3 steps.

#> Episode 69 finished after 9 steps.

#> Episode 70 finished after 13 steps.

#> Episode 71 finished after 9 steps.

#> Episode 72 finished after 11 steps.

#> Episode 73 finished after 3 steps.

#> Episode 74 finished after 7 steps.

#> Episode 75 finished after 5 steps.

#> Episode 76 finished after 9 steps.

#> Episode 77 finished after 9 steps.

#> Episode 78 finished after 15 steps.

#> Episode 79 finished after 5 steps.

#> Episode 80 finished after 19 steps.

#> Episode 81 finished after 11 steps.

#> Episode 82 finished after 5 steps.

#> Episode 83 finished after 5 steps.

#> Episode 84 finished after 5 steps.

#> Episode 85 finished after 25 steps.

#> Episode 86 finished after 5 steps.

#> Episode 87 finished after 9 steps.

#> Episode 88 finished after 5 steps.

#> Episode 89 finished after 3 steps.

#> Episode 90 finished after 7 steps.

#> Episode 91 finished after 9 steps.

#> Episode 92 finished after 19 steps.

#> Episode 93 finished after 15 steps.

#> Episode 94 finished after 9 steps.

#> Episode 95 finished after 5 steps.

#> Episode 96 finished after 3 steps.

#> Episode 97 finished after 11 steps.

#> Episode 98 finished after 3 steps.

#> Episode 99 finished after 5 steps.

#> Episode 100 finished after 3 steps.